Separate into own fcidump.py file
